Rename TestCode.js
