+++ src/containers/MealList.js
@@ -1,57 +1,92 @@
-// import axios from 'axios';
-// import React, { useState, useEffect } from 'react';
-// import { useDispatch, useSelector } from 'react-redux';
-// import Meal from '../components/Meal';
-// import { setMeals, changeFilter } from '../Redux/Actions/mealActions';
-// import Loading from '../components/Loading';
-// import CategoryFilter from '../components/CategoryFilter';
+import { useEffect, useState } from 'react';
+import axios from 'axios';
+import { useSelector, useDispatch } from 'react-redux';
+import Loading from '../components/Loading';
+import { setMeals, changeFilter } from '../Redux/Actions/mealActions';
+import Meal from '../components/Meal';
 
+const MealList = () => {
+  const myCategory = useSelector((state) => state.filter);
+  const recipes = useSelector((state) => state.recipe.myMeals);
+  const [loading, setLoading] = useState(true);
+  const dispatch = useDispatch();
+  const filters = [
+    'All',
+    'Beef',
+    'Breakfast',
+    'Chicken',
+    'Dessert',
+    'Goat',
+    'Lamb',
+    'Miscellaneous',
+    'Pasta',
+    'Pork',
+    'Seafood',
+    'Side',
+    'Starter',
+    'Vegan',
+    'Vegetarian',
+  ];
 
-// const MealList = () => {
-//   const [loading, setLoading] = useState(true);
-//    const dispatch = useDispatch();
-//   // const url = 'https://www.themealdb.com/api/json/v1/1/categories.php';
+  const myFetch = () => {
+    const result = [];
+    filters.map((myCategory) => {
+      const url = `https://www.themealdb.com/api/json/v1/1/filter.php?c=${myCategory}`;
 
-//   // const fetchMeals = async () => {
-//   //   setLoading(true);
-//   //   try {
-//   //     const response = await axios.get(url);
-//   //     dispatch(setMeals(response.data.categories));
-//   //     setLoading(false);
-//   //   } catch (error) {
-//   //     console.log(error);
-//   //     setLoading(false);
-//   //   }
-//   // };
+      axios
+        .get(url)
+        .then((response) => {
+          const ability = response.data.meals
+          const newCat = ability.map((cat) => ({
+            ...cat,
+            category: `${myCategory}`,
+          }));
+          result.push(...newCat);
+          dispatch(setMeals(result));
+          setLoading(false);
+        })
+        .catch((err) => {
+          return err;
+        });
+    });
+  };
 
-//   // useEffect(() => {
-//   //   fetchMeals();
-//   // }, []);
+  useEffect(() => {
+    myFetch();
+  }, []);
 
-//   const handleFilterChange = (e) => {
-//     e.preventDefault();
+  if (loading) {
+    return (
+      <main>
+        <Loading />;
+      </main>
+    );
+  }
 
-//     dispatch(changeFilter(e.target.value));
-//   };
+  const handleFilterChange = (e) => {
+    e.preventDefault();
 
-//   if (loading) {
-//     return (
-//       <main>
-//         <Loading />;
-//       </main>
-//     );
-//   }
+    dispatch(changeFilter(e.target.value));
+  };
+  const filteredMeals = recipes.filter(
+    (meal) => meal.category === myCategory || myCategory === 'All',
+  );
+  return (
+    <>
+      <div>
+        <select onChange={handleFilterChange} name="filter">
+          {filters.map((filter) => (
+            <option name="category" value={filter} key={filter}>
+              {filter}
+            </option>
+          ))}
+        </select>
+        <div>
+          <Meal filteredMeals={filteredMeals} />
+        </div>
+      </div>
+    </>
+  );
+};
 
-//   return (
-//     <>
-//       <div>
-//         <div>
-//           <CategoryFilter handleFilterChange={handleFilterChange} />
-//           <Meal />
-//         </div>
-//       </div>
-//     </>
-//   );
-// };
-
-// export default MealList;
+export default MealList;
